Make tests work in Plone 4.1
-  Rename acceptance to robot to align plone.app.robotframework guidelines
