+++ src/plone/app/imagecropping/testing.py
@@ -46,6 +46,6 @@ PLONE_APP_IMAGECROPPING_INTEGRATION = IntegrationTesting(
 PLONE_APP_IMAGECROPPING_FUNCTIONAL = FunctionalTesting(
     bases=(PLONE_APP_IMAGECROPPING, ),
     name="PLONE_APP_IMAGECROPPING_FUNCTIONAL")
-PLONE_APP_IMAGECROPPING_ACCEPTANCE = FunctionalTesting(
+PLONE_APP_IMAGECROPPING_ROBOT = FunctionalTesting(
     bases=(PLONE_APP_IMAGECROPPING, z2.ZSERVER_FIXTURE),
-    name="PLONE_APP_IMAGECROPPING_ACCEPTANCE")
+    name="PLONE_APP_IMAGECROPPING_ROBOT")
