+++ src/plone/app/imagecropping/tests/robot/test_imagecropping.txt
